changed: relocate the remaining files in opm/core/utility
move to opm/common/utility/numeric
