+++ opm/polymer/PolymerProperties.cpp
@@ -24,7 +24,7 @@
 #include <cmath>
 #include <iostream>
 #include <vector>
-#include <opm/core/utility/linearInterpolation.hpp>
+#include <opm/common/utility/numeric/linearInterpolation.hpp>
 #include <opm/common/ErrorMacros.hpp>
 #include <opm/common/Exceptions.hpp>
