+++ opm/polymer/TransportSolverTwophasePolymer.cpp
@@ -23,7 +23,7 @@
 #include <opm/polymer/TransportSolverTwophasePolymer.hpp>
 #include <opm/core/props/IncompPropertiesInterface.hpp>
 #include <opm/core/grid.h>
-#include <opm/core/utility/RootFinders.hpp>
+#include <opm/common/utility/numeric/RootFinders.hpp>
 #include <opm/core/utility/miscUtilities.hpp>
 #include <opm/core/pressure/tpfa/trans_tpfa.h>
 #include <opm/common/ErrorMacros.hpp>
